judged question124/p170

+++ 124/p170.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main(){
-    double n,p;
+    int n,p;
     cin>>n;
     for(int i=1;i<=n;++i){
         if(n%i==0)cout<<i<<" ";
